backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
exifread
python-dotenv
sqlalchemy
psycopg[binary]
torch
transformers
